requirements.txt
python-telegram-bot==22.1
yt-dlp==2025.6.25
httpx==0.28.1
httpcore==1.0.9
anyio==4.2.0
sniffio==1.3.1
h11==0.16.0
certifi==2025.6.15
idna==3.6
typing_extensions==4.10.0
exceptiongroup==1.2.0
python-dotenv
